refactor: optimize Input useEffect, move context hook
